+++ noodles-bam/src/async/reader/query.rs
@@ -4,7 +4,7 @@ use futures::{stream, Stream};
 use noodles_bgzf as bgzf;
 use noodles_core::region::Interval;
 use noodles_csi::index::reference_sequence::bin::Chunk;
-use noodles_sam::alignment::Record;
+use noodles_sam::{self as sam, alignment::Record};
 use tokio::io::{self, AsyncRead, AsyncSeek};
 
 use super::Reader;
@@ -22,6 +22,7 @@ where
 {
     reader: &'a mut Reader<bgzf::AsyncReader<R>>,
 
+    header: &'a sam::Header,
     chunks: vec::IntoIter<Chunk>,
 
     reference_sequence_id: usize,
@@ -30,18 +31,20 @@ where
     state: State,
 }
 
-pub fn query<R>(
-    reader: &mut Reader<bgzf::AsyncReader<R>>,
+pub fn query<'a, R>(
+    reader: &'a mut Reader<bgzf::AsyncReader<R>>,
+    header: &'a sam::Header,
     chunks: Vec<Chunk>,
     reference_sequence_id: usize,
     interval: Interval,
-) -> impl Stream<Item = io::Result<Record>> + '_
+) -> impl Stream<Item = io::Result<Record>> + 'a
 where
     R: AsyncRead + AsyncSeek + Unpin,
 {
     let ctx = Context {
         reader,
 
+        header,
         chunks: chunks.into_iter(),
 
         reference_sequence_id,
@@ -62,7 +65,7 @@ where
                         None => State::Done,
                     };
                 }
-                State::Read(chunk_end) => match next_record(ctx.reader).await? {
+                State::Read(chunk_end) => match next_record(ctx.reader, ctx.header).await? {
                     Some(record) => {
                         if ctx.reader.virtual_position() >= chunk_end {
                             ctx.state = State::Seek;
@@ -80,14 +83,20 @@ where
     }))
 }
 
-async fn next_record<R>(reader: &mut Reader<bgzf::AsyncReader<R>>) -> io::Result<Option<Record>>
+async fn next_record<R>(
+    reader: &mut Reader<bgzf::AsyncReader<R>>,
+    header: &sam::Header,
+) -> io::Result<Option<Record>>
 where
     R: AsyncRead + AsyncSeek + Unpin,
 {
     let mut record = Record::default();
 
-    reader.read_record(&mut record).await.map(|n| match n {
-        0 => None,
-        _ => Some(record),
-    })
+    reader
+        .read_record(header, &mut record)
+        .await
+        .map(|n| match n {
+            0 => None,
+            _ => Some(record),
+        })
 }